backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3
pandas==1.5.3
python-dateutil==2.8.2
pandas-market-calendars>=4.0
yfinance>=0.2.0
requests==2.31.0
python-dotenv==1.0.0
scipy>=1.9.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2
